+++ config/routes.rb
@@ -23,6 +23,8 @@ Rails.application.routes.draw do
     resources :replies, only: :index
     get :spam_inbox, path: "spam-inbox"
     get :spam_report, path: "spam-report(/:year)"
+    get :profile_review, path: "profile-review"
+    post :submit_profile_review, path: "profile-review"
     post :unspamban, path: "unspamban/:id"
     post :destroy_spammer, path: "empty-spam/:id"
     post :empty_spam, path: "empty-spam"