+++ db/migrate/20180509143815_add_profile_reviewed_to_users.rb
@@ -1,0 +1,6 @@
+class AddProfileReviewedToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :profile_reviewed, :boolean, null: false, default: false
+    add_index :users, :profile_reviewed
+  end
+end